% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bot_eft}
\alias{bot_eft}
\title{Elliptical Fourier transforms coefficients from bottles dataset}
\format{
A \code{coe_tbl} with:
\itemize{
\item \code{coe} (\code{coe_list}): eft coefficients
\item \code{type} (\code{factor}): bottle type
\item \code{fake} (\code{factor}): bottle type
}
}
\source{
Borrowed default shapes from (c) Adobe Photoshop. Please do not send me to jail.
}
\description{
Obtained with \code{bot2 \%>\% efourier(6) \%>\% dplyr::mutate(q=rnorm(40, 5)) \%>\% dplyr::select(coe, type, fake, q)}
}
\examples{
bot_eft
}
\concept{outlines}
