% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_lda_gr.R
\name{gg_CV}
\alias{gg_CV}
\alias{gg_CV.default}
\alias{gg_CV.stat_lda_full}
\alias{gg_CV.matrix}
\title{Plot confusion matrices}
\usage{
gg_CV(
  x,
  prop = TRUE,
  percent = TRUE,
  drop_zeros = TRUE,
  text = TRUE,
  text_colour = "white",
  text_size = 8,
  text_signif = 2,
  axis_size = 10,
  axis_x_angle = 0
)

\method{gg_CV}{default}(
  x,
  prop = TRUE,
  percent = TRUE,
  drop_zeros = TRUE,
  text = TRUE,
  text_colour = "white",
  text_size = 8,
  text_signif = 2,
  axis_size = 10,
  axis_x_angle = 0
)

\method{gg_CV}{stat_lda_full}(
  x,
  prop = TRUE,
  percent = TRUE,
  drop_zeros = TRUE,
  text = TRUE,
  text_colour = "white",
  text_size = 8,
  text_signif = 2,
  axis_size = 10,
  axis_x_angle = 0
)

\method{gg_CV}{matrix}(
  x,
  prop = TRUE,
  percent = TRUE,
  drop_zeros = TRUE,
  text = TRUE,
  text_colour = "white",
  text_size = 8,
  text_signif = 2,
  axis_size = 10,
  axis_x_angle = 0
)
}
\arguments{
\item{x}{confusion matrix or \code{CV_tbl} returned by lda}

\item{prop}{\code{logical} if \code{TRUE} (default), use accuracies (within an actual class); if \code{FALSE} use counts}

\item{percent}{\code{logical} whether to use percentages (default to \code{TRUE})}

\item{drop_zeros}{\code{logical} whether to drop zeros in labels (default to \code{TRUE})}

\item{text}{\code{logical} whether display cell information (default to \code{TRUE})}

\item{text_colour}{\code{logical} whether display cell information (default to \code{white})}

\item{text_size}{\code{numeric} size for cell information, in points (default to \code{8})}

\item{text_signif}{\code{integer} to feed \link{signif} (default to `2``)}

\item{axis_size}{\code{numeric} size for axes labels}

\item{axis_x_angle}{\code{numeric} angle for x-axis labels}
}
\description{
Plots as heatmap and with useful cosmetics
}
\details{
when a \code{matrix} is passed, it assumes that it is:
\itemize{
\item squared
\item rows are for actual classes
\item columns for predicted classes
\item class accuracies are calculated per actual classes, ie rowwise
}
}
\examples{
# works fine on matrices too
# set.seed(2329) # for the sake of replicability
m <- matrix(sample(1:100), 10)
gg_CV(m)

}
