% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momstats-package.R
\name{Momocs2-reexports}
\alias{Momocs2-reexports}
\alias{fold}
\alias{unfold}
\alias{get_coe_cols}
\title{Functions re-exported from Momocs2}
\description{
These functions are re-exported from Momocs2 for convenient access in Momstats.
See the Momocs2 documentation for details.
}
\keyword{internal}
