% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Momstats-package.R
\docType{package}
\name{Momstats-package}
\alias{Momstats}
\alias{Momstats-package}
\title{Momstats: Statistical Methods for Morphometric Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The goal of 'Momstats' is to provide statistical methods on top of 'Momocs2' tibbles, and chiefly multivariate methods on morphometric data. This includes mean shapes, principal component analyses along with morphological space reconstruction, multivariate hypothesis testing, linear discriminant analyses with permutations and balancing, hierachical clustering methods, kmeans and kmedoids, etc. Visualization is at heart with several extensions to 'ggplot2' graphics. A light dispatching strategy makes it comfortable and consistent to use.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MomX/Momstats}
  \item \url{https://momx.github.io/Momstats/}
  \item Report bugs at \url{https://github.com/MomX/Momstats/issues}
}

}
\author{
\strong{Maintainer}: Vincent Bonhomme \email{bonhomme.vincent@gmail.com} (\href{https://orcid.org/0000-0002-2742-6349}{ORCID})

}
\keyword{internal}
