% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_lda.R
\name{stat_lda}
\alias{stat_lda}
\alias{stat_lda0}
\title{Linear discriminant analysis}
\usage{
stat_lda0(x, f, full = FALSE, ...)

stat_lda(x, f, ...)
}
\arguments{
\item{x}{\verb{matrix or [tibble][tibble::tibble-package] containing the explanatory variables forwarded to [MASS::lda]'s }x`}

\item{f}{\code{factor} forwarded to \link[MASS:lda]{MASS::lda}'s \code{grouping}. If missing takes first columns and remove it from x.}

\item{full}{\code{logical} whether to prepare useful components (default to \code{FALSE})}

\item{...}{addtional parameters forwardes to \link[MASS:lda]{MASS::lda}}
}
\description{
Calculations are delegated to \link[MASS:lda]{MASS::lda}
}
\details{
With \code{full=TRUE}, it is at least 6 times slower.
}
\section{Functions}{
\itemize{
\item \code{stat_lda0}: Vanilla lda

\item \code{stat_lda}: Wrapped lda
}}

\examples{
(x <- dummy_df \%>\% stat_lda_prepare(foo2_NA, a:e))

stat_lda0(x$coe_naked, x$f_naked)

stat_lda(dummy_df, foo2_NA, a:e)
}
\concept{lda stats}
