% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_lda.R
\name{stat_lda0}
\alias{stat_lda0}
\title{Linear discriminant analysis}
\usage{
stat_lda0(x, f, full = FALSE)
}
\arguments{
\item{x}{\verb{matrix or [tibble][tibble::tibble-package] containing the explanatory variables forwarded to [MASS::lda]'s }x`}

\item{f}{\code{factor} forwarded to \link[MASS:lda]{MASS::lda}'s \code{grouping}}

\item{full}{\code{logical} whether to prepare useful components (default to \code{FALSE})}
}
\description{
Calculations are delegated to \link[MASS:lda]{MASS::lda}
}
\details{
With \code{full=TRUE}, it is at least 6 times slower.
}
\seealso{
Other lda: 
\code{\link{stat_lda_prepare}()}
}
\concept{lda}
