% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_lda.R
\name{stat_lda_prepare}
\alias{stat_lda_prepare}
\title{Linear discriminant analysis preparation}
\usage{
stat_lda_prepare(
  x,
  f,
  ...,
  constant_below_var = 1e-05,
  collinear_above_cor = 1 - 1e-05
)
}
\arguments{
\item{x}{\link[tibble:tibble-package]{tibble}, typically a \link{coe_tbl}}

\item{f}{column specifying grouping factor}

\item{...}{columns specifying those to use}

\item{constant_below_var}{\code{numeric} threshold under which columns
will be considered constant (using \link[stats:var]{stats::var}) and dropped (default to \code{1e-5})}

\item{collinear_above_cor}{\code{numeric} threshold above which columns
will be considered collinear (using \link[stats:cor]{stats::cor}), and dropped (default to \code{1-1e-5})}
}
\description{
Removes NAs, constant and collinear columns.
Useful to speed up calculations when used with \link{stat_lda0}
}
\seealso{
Other lda: 
\code{\link{stat_lda0}()}
}
\concept{lda}
