% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_manova.R
\name{stat_manova}
\alias{stat_manova}
\alias{stat_manova_pairwise}
\title{Multivariate analysis of variance}
\usage{
stat_manova(x, f, ..., test = "Pillai")

stat_manova_pairwise(x, f, ..., test = "Pillai")
}
\arguments{
\item{x}{\link[tibble:tibble-package]{tibble}}

\item{f}{\code{factor} used as right hand side in \link[stats:manova]{stats::manova}'s formula}

\item{...}{columns used as left hand side in \link[stats:manova]{stats::manova}'s formula}

\item{test}{forwarded to \link[broom:tidy.manova]{broom::tidy.manova}'s \code{test}. One of "Pillai" (Pillai's trace, default),
"Wilks" (Wilk's lambda), "Hotelling-Lawley" (Hotelling-Lawley trace) or
"Roy" (Roy's greatest root) indicating which test statistic should be used.}
}
\value{
see \link[broom:tidy.manova]{broom::tidy.manova}'s Value section
}
\description{
Calculations are delegated to \link[stats:manova]{stats::manova}
}
\section{Functions}{
\itemize{
\item \code{stat_manova_pairwise}: pairwise manova for every pair of levels
}}

\concept{manova}
