% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vitis}
\alias{vitis}
\title{Grapevine pips outlines (two views) with elliptic Fourier transform coefficients}
\format{
A tibble with 410 rows and 5 variables:
\itemize{
\item \code{id}: for the accession
\item \code{pip}: Numeric identifying pip number
\item \code{accession}: Character 14 names
\item \code{status}: Factor with 2 levels (domesticated/wild)
\item \code{length} : pip length measured before outline normalization
\item \code{VD}: 5 eft harmonics for dorsal view
\item \code{VL}: 5 eft harmonics for lateral view
}
}
\source{
adapted from Bonhomme, Vincent; Ivorra, Sarah; Figueiral, Isabel; Pastor, Thierry; Terral, Jean-Frédéric; Bouby, Laurent (2021).
Vitis reference collection database (Sci Reports 2021). figshare. Dataset. https://doi.org/10.6084/m9.figshare.14170484.v1
}
\usage{
vitis
}
\description{
A dataset containing 410 pip outlines (dorsal and lateral view) that have been preprocessed and
transformed using elliptic Fourier analysis (5 harmonics). This dataset is used for
demonstrating statistical methods in Momstats.
}
\examples{
# View structure
vitis

}
\keyword{datasets}
