% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_pca.R
\name{plot.stat_pca}
\alias{plot.stat_pca}
\title{Plot PCA results}
\usage{
\method{plot}{stat_pca}(
  x,
  type = c("scores", "scree", "loadings"),
  pcs = c(1, 2),
  extent = 1,
  morphospace = "grid",
  color = NULL,
  labels = NULL,
  chull = TRUE,
  n_pcs = NULL,
  cex = 0.7,
  legend = TRUE,
  ...
)
}
\arguments{
\item{x}{A \code{stat_pca} object}

\item{type}{Character. Type of plot}

\item{pcs}{Integer vector of length 2. Which PCs to plot}

\item{extent}{Numeric. Plot extent factor (< 1 = zoom in, > 1 = zoom out)}

\item{morphospace}{Character or FALSE. Morphospace type: "grid", "range", "axes", "centroids", FALSE}

\item{color}{Column name for coloring points}

\item{labels}{Column name for text labels}

\item{chull}{Logical. Draw convex hulls?}

\item{n_pcs}{Integer. For scree plot}

\item{cex}{Numeric. Character expansion}

\item{legend}{Logical. Show legend?}

\item{...}{Passed to morphospace functions (n, nrow, ncol, template, lwd, col)}
}
\description{
Plot PCA results
}
